app/requirements.txt
numpy==2.1.2
opencv-python-headless==4.10.0.84
tqdm==4.66.5
faiss-cpu==1.9.0
insightface==0.7.3
onnxruntime==1.19.2
streamlit==1.39.0
pillow==10.4.0

